rename simple get stack file to momento lambda stack
