+++ examples/aws-lambda/infrastructure/lib/momento-lambda-stack.ts
